added savePlace function to PlacesList and pass place props to it via favoritePlace function on PlaceTile
