+++ client/src/components/PlaceTile.js
@@ -7,49 +7,13 @@ import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
 const PlaceTile = (props) => {
   const { distance, id, image, location, name, price, rating, url } = props
   const getLocation = location.join(' ')
-  const [place, setPlace] = useState({
-    distance: "",
-    id: "",
-    image: "",
-    location: "",
-    name: "",
-    price: "",
-    rating: "",
-    url: ""
-  })
+  const [place, setPlace] = useState({})
 
-  const savePlace = async (placePayload) => {
-    debugger
-    try {
-      debugger
-      const response = await fetch("/api/v1/places", {
-        method: "POST",
-        headers: new Headers({
-          "Content-Type": "application/json"
-        }),
-        body: JSON.stringify(placePayload)
-      })
-      if (!response.ok) {
-        if (response.status === 422) {
-          const body = await response.json()
-          const newErrors = translateServerErrors(body.errors)
-          return setErrors(newErrors)
-        } else {
-          const errorMessage = `${response.status} (${response.statusText})`
-          const error = new Error(errorMessage)
-          throw error
-        }
-      } else {
-      const body = await response.json()
-      setPlace(body)
-      debugger
-      }
-    } catch (error) {
-      console.error(`Error in fetch: ${error.message}`)
-    }
+  const favoritePlace = event => {
+    event.preventDefault()
+    props.savePlace(props)
   }
   
-  //each time you click add the favorites you concatinate then add to state 
   return (
     <div className="cell">
       <div className="card">
@@ -71,7 +35,8 @@ const PlaceTile = (props) => {
         <FavoriteBorderIcon 
         className="fav-icon" 
         fontSize="large"
-        onClick={savePlace}
+        place={place}
+        onClick={favoritePlace}
         />
         </div>
       </div>
